+++ sdk/brcodePreview/index.js
@@ -1,0 +1,4 @@
+const brcodePreview = require('./brcodePreview.js');
+
+exports.query = brcodePreview.query;
+exports.BrcodePreview = brcodePreview.BrcodePreview;