+++ sdk/deposit/index.js
@@ -1,0 +1,6 @@
+const deposit = require('./deposit.js');
+
+exports.log = require('./log');
+exports.get = deposit.get;
+exports.query = deposit.query;
+exports.Deposit = deposit.Deposit;