+++ sdk/dictKey/index.js
@@ -1,0 +1,5 @@
+const dictKey = require('./dictKey.js');
+
+exports.get = dictKey.get;
+exports.query = dictKey.query;
+exports.DictKey = dictKey.DictKey;