+++ sdk/transfer/transfer.js
@@ -16,13 +16,13 @@ class Transfer extends Resource {
      * @param amount [integer]: amount in cents to be transferred. ex: 1234 (= R$ 12.34)
      * @param name [string]: receiver full name. ex: 'Anthony Edward Stark'
      * @param taxId [string]: receiver tax ID (CPF or CNPJ) with or without formatting. ex: '01234567890' or '20.018.183/0001-80'
-     * @param bankCode [string]: 1 to 3 digits of the receiver bank institution in Brazil. ex: '200' or '341'
+     * @param bankCode [string]: code of the receiver bank institution in Brazil. If an ISPB (8 digits) is informed, a PIX transfer will be created, else a TED will be issued. ex: '20018183' or '341'
      * @param branchCode [string]: receiver bank account branch. Use '-' in case there is a verifier digit. ex: '1357-9'
      * @param accountNumber [string]: Receiver Bank Account number. Use '-' before the verifier digit. ex: '876543-2'
      *
      * Parameters (optional):
      * @param tags [list of strings]: list of strings for reference when searching for transfers. ex: ['employees', 'monthly']
-     * @param scheduled [string, default now]: datetime when the transfer will be processed. May be pushed to next business day if necessary. ex: '2020-03-11 08:00:00.000'
+     * @param scheduled [string, default now]: date or datetime when the transfer will be processed. May be pushed to next business day if necessary. ex: '2020-11-12T00:14:22.806+00:00' or '2020-11-30'
      *
      * Attributes (return-only):
      * @param id [string, default null]: unique id returned when Transfer is created. ex: '5656565656565656'
@@ -151,7 +151,7 @@ exports.query = async function ({ limit, after, before, transactionIds, status, 
      * @param before [string, default null]: date filter for objects created or updated only before specified date. ex: '2020-03-10'
      * @param transactionIds [list of strings, default null]: list of transaction IDs linked to the desired transfers. ex: ['5656565656565656', '4545454545454545']
      * @param status [string, default null]: filter for status of retrieved objects. ex: 'success' or 'failed'
-     * @param taxId [string, default null]: filter for transfers sent to the specified tax ID.ex: "012.345.678-90"
+     * @param taxId [string, default null]: filter for transfers sent to the specified tax ID.ex: '012.345.678-90'
      * @param sort [string, default '-created']: sort order considered in response. Valid options are 'created', '-created', 'updated' or '-updated'.
      * @param tags [list of strings, default null]: tags to filter retrieved objects. ex: ['tony', 'stark']
      * @param ids [list of strings, default null]: list of ids to filter retrieved objects. ex: ['5656565656565656', '4545454545454545']